Update test for AW
